Vspei 638 asignar tarjetas de forma masiva por excel a usuarios [Backend] (#437)
Se realiza lo siguiente:
- Se crea enpoint para asignar tarjetas desde archivo excel.
- Se corrige bug donde para realizar una transferencia donde ya permite recibir centavos.

User History [Jira]: VSPEI-638, VSPEI-639,  VSPEI-663

+++ src/stp/cardCloud/application/create_card_transfer/CardCloudCardTransferCreator.php
@@ -17,7 +17,7 @@ final class CardCloudCardTransferCreator
         string $source,
         string $destinationType,
         string $destination,
-        int    $amount,
+        float  $amount,
         string $description
     ): CardCloudServiceResponse
     {
@@ -30,7 +30,7 @@ final class CardCloudCardTransferCreator
             'description' => $description
         ];
         return new CardCloudServiceResponse(
-            $this->repository->createTransfer($businessId,$transferData)
+            $this->repository->createTransfer($businessId, $transferData)
         );
     }
 }
